WORKING MOTOR CONTROLLER WITH PEDALS

+++ software/Core/Inc/CC_FSM_States.h
@@ -68,14 +68,14 @@ typedef struct
 	uint32_t brakeAdcValues[100];
 	uint32_t accelAdcValues[150];
 	uint32_t brakePressureThreshold;
-	uint16_t rollingBrakeValues[10]; uint16_t secondaryRollingBrakeValues[10];
-	uint16_t brakeMin[2]; uint16_t brakeMax[2];
-	uint16_t accelMin[3]; uint16_t accelMax[3];
+	uint32_t rollingBrakeValues[10]; uint32_t secondaryRollingBrakeValues[10];
+	uint32_t brakeMin[2]; uint32_t brakeMax[2];
+	uint32_t accelMin[3]; uint32_t accelMax[3];
 	uint32_t rollingAccelValues[10]; uint32_t secondaryRollingAccelValues[10]; uint32_t tertiaryRollingAccelValues[10];
 
 	/* Formatted Pedal Travel Positions */
-	uint8_t accelTravel;
-	uint8_t brakeTravel;
+	int32_t accelTravel;
+	int32_t brakeTravel;
 
 	bool tractiveActive;
 	bool faultDetected;
